+++ src/main/java/shop/tryit/repository/payment/PaymentJpaRepository.java
@@ -2,7 +2,7 @@ package shop.tryit.repository.payment;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.tryit.domain.payment.Payment;
+import shop.tryit.domain.payment.entity.Payment;
 
 public interface PaymentJpaRepository extends JpaRepository<Payment, Long> {
     Optional<Payment> findByNumber(Long number);
